+++ packages/chains/webpack.config.js
