+++ packages/connectors/package.json
@@ -43,7 +43,7 @@
     "lint": "eslint -c '../../.eslintrc' --fix './src/**/*.ts'"
   },
   "dependencies": {
-    "@web3modal/chain-utils": "^1.0.0"
+    "@web3modal/chains": "^1.0.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.12.1",
