+++ packages/constants/src/index.ts
@@ -1,4 +1,3 @@
-export * from "./chains";
 export * from "./classNames";
 export * from "./events";
 export * from "./keys";
